+++ src/services/TeenPatti/TeenPattiGame.js
@@ -19,12 +19,24 @@ class TeenPattiGame extends BaseGame {
     this.BETTING_PHASE_DURATION = 30000; // 30 seconds for betting
     this.CARD_DEAL_DURATION = 5000; // 5 seconds for dealing
     this.gameInterval = null;
+    this.betSides = ["playerA", "playerB"];
+  }
+
+  collectCards(playerSide) {
+    switch (playerSide) {
+        case "A":
+            return this.bettingResults.player1;
+        case "B":
+            return this.bettingResults.player2;
+        default:
+            return [];
+    }
   }
 
   logSpecificGameState() {
-    // console.log("Blind Card:", this.blindCard);
-    // console.log("Player 1 Cards:", this.player1Cards);
-    // console.log("Player 2 Cards:", this.player2Cards);
+    console.log("Blind Card:", this.blindCard);
+    console.log("Player 1 Cards:", this.player1Cards);
+    console.log("Player 2 Cards:", this.player2Cards);
   }
 
   async saveState() {
@@ -51,7 +63,7 @@ class TeenPattiGame extends BaseGame {
 
   async recoverState() {
     try {
-      await super.recoverState();
+      await this.recoverState();
       const state = await redis.hgetall(`game:${this.gameId}:teenpatti`);
       if (state && Object.keys(state).length) {
         this.blindCard = state.blindCard ? JSON.parse(state.blindCard) : null;
@@ -77,7 +89,7 @@ class TeenPattiGame extends BaseGame {
   async start() {
     this.status = GAME_STATES.BETTING;
     this.startTime = Date.now();
-    await this.saveState();
+    await super.saveState();
 
     this.logGameState("Game Started - Betting Phase");
 
@@ -102,7 +114,7 @@ class TeenPattiGame extends BaseGame {
       this.deck.shift(),
     ];
 
-    await this.saveState();
+    await super.saveState();
     this.logGameState("Dealing Phase Started");
 
     setTimeout(async () => {
@@ -114,7 +126,7 @@ class TeenPattiGame extends BaseGame {
     const result = compareHands(this.player1Cards, this.player2Cards);
     this.status = GAME_STATES.COMPLETED;
     this.winner = result === 1 ? "player1" : "player2";
-    await this.saveState();
+    await super.saveState();
 
     this.logGameState("Winner Determined");
 
@@ -196,7 +208,7 @@ class TeenPattiGame extends BaseGame {
 
   async endGame() {
     this.status = GAME_STATES.COMPLETED;
-    await this.saveState();
+    await super.saveState();
     await this.storeGameResult();
 
     this.logGameState("Game Completed");
